TRIVIAL: Change locator at analyze page: s/category/categories/
For consistency we are changing bucket naming to plural. This will land in analytical designer with the undo/redo feature and this is a suggested change to the test code. Please note that I haven't had the opportunity to try it out, as I don't have the environment set up on my machine.