FEATURE: Integration of UI tests with browserstack
Thanks for
https://github.com/MatousJobanek/arquillian-drone-browserstack-webdriver
we are able to integrate the tests easily from arquillian.xml (using the
new extension which is now part of this repository - it will be changed
to external dependency when it's done).

There is a new property:
<!-- possible values are:
local/browserstack-mobile/browserstack-platform -->
`<test.execution.env>local</test.execution.env>`

By default, you can always run the tests locally (arquillian.xml), but
if you want to run the tests in browserstack, you just switch the
test.execution.env to browserstack-mobile/browserstack-platform and
provide required parameters at new arquillian-browserstack-*.xml.

Next step is to improve the hanlind of specific devices/configurations.