TRIVIAL: Support for skipping validations
Do not run validations on any child of AbstractProjectTest.

Boolean property is defined as part of the BeforeClass in child tests,
enabled by default to keep the current status quo.

It was switched only on the indigo dashboards for now.

Another option is to move this ability in TestParams and update all
graphene jobs to have this option on/off.